Remove HTML comments an empty lines
Note: It looks like grunt-replace will not iterate over the file
for each match. For example, blank lines left by HTML comments
that are removed will not be match and removed by the empty line
pattern.

For comments in particular, I could run grunt-replace twice--
or just comment as ejs in the first place.